refactor: split descriptor and inspector
